feat: Add SQL validator to composer.json
This commit introduces the SQL validator "codelicia/xulieta-sql" as
a suggested package in the composer.json file. By including this
validator, the project gains the capability to validate SQL code in
addition to the existing JSON and PHP Litterate Programming validators.

This feature enhancement expands the range of validation options
available for the project, contributing to improved code quality and
ensuring adherence to SQL standards.

Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>